refactor(terraform): move provider specific logic `provider.ts`
