+++ plugins/terraform/init.ts
@@ -6,7 +6,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-import { TerraformProvider } from "."
+import { TerraformProvider } from "./provider"
 import { applyStack, getRoot, getStackStatus, getTfOutputs, prepareVariables, setWorkspace } from "./common"
 import chalk from "chalk"
 import { deline } from "@garden-io/sdk/util/string"
