+++ plugins/terraform/test/terraform.ts
@@ -15,7 +15,8 @@ import { getRootLogMessages, makeTestGarden, TestGarden } from "@garden-io/sdk/t
 import { findByName } from "@garden-io/core/build/src/util/util"
 import { getTerraformCommands } from "../commands"
 import { ConfigGraph, LogLevel } from "@garden-io/sdk/types"
-import { gardenPlugin, TerraformProvider } from ".."
+import { gardenPlugin } from ".."
+import { TerraformProvider } from "../provider"
 import { DeployTask } from "@garden-io/core/build/src/tasks/deploy"
 import { getWorkspaces, setWorkspace } from "../common"
 import { resolveAction } from "@garden-io/core/build/src/graph/actions"
